requirements.txt
click==8.0.4
Flask==2.0.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
psycopg2-binary
requests
python-dateutil
twilio
vobject